tests/wpt/mozilla/tests/mozilla/mozbrowser/mozbrowsericonchange_event.html

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(t) {

  var expectedEvents = [
    {rel: "icon", href: "http://web-platform.test:8000/test1.ico", sizes: "any"},
    {rel: "icon", href: "http://web-platform.test:8000/_mozilla/mozilla/mozbrowser/test2.ico", sizes: "16x16 24x24"},
    {rel: "shortcut icon", href: "http://example.com/test3.ico", sizes: ""},
    {rel: "apple-touch-icon", href: "http://web-platform.test:8000/test4.ico", sizes: ""},
    {rel: "icon", href: "http://web-platform.test:8000/test5.ico", sizes: ""},
    {rel: "icon", href: "http://web-platform.test:8000/test6.ico", sizes: "any"},
    {rel: "icon", href: "http://web-platform.test:8000/test6.ico", sizes: "32x32"},
  ];

  var receivedEvents = [];

  var iframe = document.createElement("iframe");
  iframe.mozbrowser = "true";
  iframe.src = "mozbrowsericonchange_event_test.html";

  iframe.addEventListener("mozbrowsericonchange", t.step_func(e => {

    receivedEvents.push({
      rel: e.detail.rel,
      href: e.detail.href,
      sizes: e.detail.sizes,
    });

    if (receivedEvents.length == expectedEvents.length) {
      for (var i = 0; i < expectedEvents.length; i++) {
        var e1 = expectedEvents[i];
        var e2 = receivedEvents[i];
        assert_equals(e1.rel, e2.rel);
        assert_equals(e1.href, e2.href);
        assert_equals(e1.sizes, e2.sizes);
      }
      t.done();
    }
  }));


  document.body.appendChild(iframe);
});

</script>
</body>
